pyproject.toml
[tool.poetry]
name = "aia-language-speech-reconigtion"
version = "0.1.1"
description = ""
authors = ["Edgar Rodrigo Sanchez De Gracia <edgr.sanchez@mercadolibre.cl>"]
readme = "README.md"
packages = [{include = "aia_language_speech_reconigtion"}]

[tool.poetry.dependencies]
python = ">=3.11.4,<3.12"
colorama = "^0.4.5"
confluent-kafka = "2.3.0"
scikit-learn = "1.3.2"
pandas = "2.1.2"
PyAudio = "0.2.13"
coverage = "^7.6.0"
pocketsphinx = "^5.0.3"
aia_utils = "^0.2.1"
SpeechRecognition = "^3.10.4"

[tool.poetry.dev-dependencies]
pytest = "^7.1.3"
pytest-dotenv = "^0.5.2"
pytest-cov = "^4.0.0"

[tool.poetry.scripts]
daemon = 'aia_language_speech_reconigtion.daemon:run'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
